python/requirements.txt
yfinance==1.1.0
supabase==2.27.3
python-dotenv==1.2.1
pandas==2.2.3
scipy==1.15.1